// frontend/apps/thunder-develop/src/features/flows/models/base.ts
export interface StrictBase<T = unknown> {
  /**
   * ID of the component or the primitive.
   */
  id: string;
  /**
   * Category of the component or the primitive.
   */
  category: string;
  /**
   * Type of the component or the primitive.
   */
  type: string;
  /**
   * Version of the component or the primitive.
   */
  version: string;
  /**
   * Is the component or the primitive deprecated.
   */
  deprecated: boolean;
  /**
   * Is the component or the primitive  deletable.
   */
  deletable: boolean;
  /**
   * Display properties of the component or the primitive.
   */
  display: BaseDisplay;
  /**
   * Configuration of the component or the primitive.
   */
  config: BaseConfig & T;
  /**
   * Base variant of the component or the primitive
   */
  variant?: unknown;
  /**
   * Variants of the component or the primitive.
   */
  variants?: Base<T>[];
  /**
   * Data added to the component by the flow builder.
   */
  data?: unknown;
}

/**
 * Interface representing a base component or a primitive.
 */
export interface Base<T = unknown> extends StrictBase<T> {
  /**
   * Type of the resource needed for visual editor operations.
   * @remarks This is a display only meta field and not being published to the backend.
   */
  resourceType: string;
}

export interface BaseDisplay {
  /**
   * Fallback & i18n key value of the label.
   */
  label: string;
  /**
   * Image URL of the component or the primitive.
   */
  image: string;
  /**
   * The default variant of the component or the primitive.
   */
  defaultVariant?: string;
  /**
   * Description of the component or the primitive.
   */
  description?: string;
  /**
   * Should the component be shown on the resource panel.
   */
  showOnResourcePanel: boolean;
}
// ...
export interface BaseConfig {
  /**
   * Field properties.
   */
  field: Field;
  /**
   * Styles of the component or the primitive.
   */
  styles: CSSProperties;
  /**
   * Identifier of the component or the primitive.
   */
  identifier?: string;
  /**
   * Label of the component or the primitive.
   */
  label?: string;
  /**
   * Placeholder of the component or the primitive.
   */
  placeholder?: string;
  /**
   * Hint of the component or the primitive.
   */
  hint?: string;
  /**
   * Is the component or the primitive required.
   */
  required?: boolean;
  /**
   * Should the password field require confirmation.
   */
  requireConfirmation?: boolean;
}